+++ packages/common/CHANGELOG.md
@@ -1,3 +1,10 @@
+## @esri/hub-common [20.4.5](https://github.com/Esri/hub.js/compare/@esri/hub-common@20.4.4...@esri/hub-common@20.4.5) (2025-09-10)
+
+
+### Bug Fixes
+
+* user settings schema tolerant of misconfigured trustedOrgs ([#1994](https://github.com/Esri/hub.js/issues/1994)) ([8ba01f9](https://github.com/Esri/hub.js/commit/8ba01f92c8bbdf5256da5b5afc3e2edfb18bfebf))
+
 ## @esri/hub-common [20.4.4](https://github.com/Esri/hub.js/compare/@esri/hub-common@20.4.3...@esri/hub-common@20.4.4) (2025-09-04)